requirements.txt
# Data Collection & Processing
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
schedule==1.2.0

# Machine Learning
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
folium==0.15.1

# Web Application
streamlit==1.29.0

# Utilities
numpy==1.26.2
pytz==2023.3
